requirements.txt
# generated from manifests external_dependencies
cryptography<37
extract_msg
lottie
premailer
python-telegram-bot
requests_toolbelt
